# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AbrirCaixaRequest:
    properties:
      saldo_inicial:
        type: string
    required:
    - saldo_inicial
    type: object
...
  dto.CaixaDiarioResponse:
    properties:
      created_at:
        type: string
      data_abertura:
        type: string
      data_fechamento:
        type: string
      divergencia:
        type: string
      id:
        type: string
      justificativa_divergencia:
        type: string
      operacoes:
        items:
          $ref: '#/definitions/dto.OperacaoCaixaResponse'
        type: array
      saldo_esperado:
        type: string
      saldo_inicial:
        type: string
      saldo_real:
        type: string
      status:
        type: string
      total_entradas:
        type: string
      total_reforcos:
        type: string
      total_saidas:
        type: string
      total_sangrias:
        type: string
      updated_at:
        type: string
      usuario_abertura_id:
        type: string
      usuario_abertura_nome:
        type: string
      usuario_fechamento_id:
        type: string
      usuario_fechamento_nome:
        type: string
    type: object
  dto.CaixaDiarioResumoResponse:
    properties:
      data_abertura:
        type: string
      data_fechamento:
        type: string
      divergencia:
        type: string
      id:
        type: string
      saldo_esperado:
        type: string
      saldo_inicial:
        type: string
      saldo_real:
        type: string
      status:
        type: string
      tem_divergencia:
        type: boolean
      usuario_abertura_nome:
        type: string
      usuario_fechamento_nome:
        type: string
    type: object
...
  dto.FecharCaixaRequest:
    properties:
      justificativa:
        type: string
      saldo_real:
        type: string
    required:
    - saldo_real
    type: object
...
  dto.ListCaixaHistoricoResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CaixaDiarioResumoResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.OperacaoCaixaResponse:
    properties:
      created_at:
        type: string
      descricao:
        type: string
      destino:
        type: string
      id:
        type: string
      origem:
        type: string
      tipo:
        type: string
      usuario_id:
        type: string
      usuario_nome:
        type: string
      valor:
        type: string
    type: object
...
  dto.ReforcoRequest:
    properties:
      descricao:
        minLength: 5
        type: string
      origem:
        enum:
        - TROCO
        - CAPITAL_GIRO
        - TRANSFERENCIA
        - OUTROS
        type: string
      valor:
        type: string
    required:
    - descricao
    - origem
    - valor
    type: object
...
  dto.SangriaRequest:
    properties:
      descricao:
        minLength: 5
        type: string
      destino:
        enum:
        - DEPOSITO
        - PAGAMENTO
        - COFRE
        - OUTROS
        type: string
      valor:
        type: string
    required:
    - descricao
    - destino
    - valor
    type: object
...
  dto.TotaisCaixaResponse:
    properties:
      saldo_atual:
        type: string
      total_despesas:
        type: string
      total_reforcos:
        type: string
      total_sangrias:
        type: string
      total_vendas:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nexo.barber
    name: API Support
    url: https://nexo.barber/support
  description: API completa para gestão de barbearias com módulos de Agendamento,
    Comanda, Financeiro, Estoque e Analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://nexo.barber/terms
  title: NEXO - Barber Analytics API
  version: "2.0"
paths:
...
  /api/v1/caixa/{id}:
    get:
      description: Retorna um caixa específico com operações
      parameters:
      - description: ID do caixa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CaixaDiarioResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Buscar caixa por ID
      tags:
      - Caixa
  /api/v1/caixa/aberto:
    get:
      description: Retorna o caixa aberto com todas as operações
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CaixaDiarioResponse'
        "404":
          description: Nenhum caixa aberto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Caixa aberto
      tags:
      - Caixa
  /api/v1/caixa/abrir:
    post:
      consumes:
      - application/json
      description: Abre um novo caixa diário com saldo inicial
      parameters:
      - description: Dados de abertura
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AbrirCaixaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CaixaDiarioResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Já existe caixa aberto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Abrir caixa
      tags:
      - Caixa
  /api/v1/caixa/fechar:
    post:
      consumes:
      - application/json
      description: Fecha o caixa com conferência de saldo real
      parameters:
      - description: Dados de fechamento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FecharCaixaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CaixaDiarioResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Caixa não aberto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fechar caixa
      tags:
      - Caixa
  /api/v1/caixa/historico:
    get:
      description: Lista caixas fechados com paginação
      parameters:
      - description: Data início (YYYY-MM-DD)
        in: query
        name: data_inicio
        type: string
      - description: Data fim (YYYY-MM-DD)
        in: query
        name: data_fim
        type: string
      - description: ID do usuário
        in: query
        name: usuario_id
        type: string
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Itens por página
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListCaixaHistoricoResponse'
      summary: Histórico de caixas
      tags:
      - Caixa
  /api/v1/caixa/reforco:
    post:
      consumes:
      - application/json
      description: Registra uma adição de dinheiro ao caixa
      parameters:
      - description: Dados do reforço
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReforcoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.OperacaoCaixaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Caixa não aberto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar reforço
      tags:
      - Caixa
  /api/v1/caixa/sangria:
    post:
      consumes:
      - application/json
      description: Registra uma retirada de dinheiro do caixa
      parameters:
      - description: Dados da sangria
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SangriaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.OperacaoCaixaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Caixa não aberto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar sangria
      tags:
      - Caixa
...
  /api/v1/caixa/totais:
    get:
      description: Retorna os totais por tipo de operação
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TotaisCaixaResponse'
        "404":
          description: Caixa não aberto
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Totais do caixa
      tags:
      - Caixa
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token JWT (formato: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...